commit teste para validacao dos especialistas

+++ models/ProductModel.js
@@ -8,6 +8,8 @@ const getAll = async () => {
 };
 
 const getById = async (id) => {
+  if (!id) return false;
+
   const query = 'SELECT * FROM StoreManager.products WHERE id = ?';
   const [result] = await connection.execute(query, [id]);
 
@@ -34,6 +36,8 @@ const postProduct = async (name, quantity) => {
 };
 
 const putProduct = async (id, name, quantity) => {
+  if (!id) return false;
+
 const query = 'UPDATE StoreManager.products SET name = ?, quantity = ? WHERE id = ? ';
 
 await connection.execute(query, [name, quantity, id]);
@@ -42,6 +46,8 @@ return getById(id);
 };
 
 const deleteProduct = async (id) => {
+  if (!id) return false;
+
   const query = 'DELETE FROM StoreManager.products WHERE id = ?';
   await connection.execute(query, [id]);
 };
