+++ models/SalesModel.js
@@ -14,6 +14,8 @@ ORDER BY saleId AND productId
 };
 
 const getById = async (saleId) => {
+  if (!saleId) return false;
+
   const query = `
   SELECT
   s.date 'date',
@@ -44,6 +46,8 @@ const postSale = async (soldItems) => {
 };
 
 const putSale = async (saleId, body) => {
+  if (!saleId) return false;
+
   const { productId, quantity } = body;
   const query = `UPDATE StoreManager
     .sales_products SET product_id = ?, quantity = ? WHERE sale_id = ?`;
@@ -54,6 +58,8 @@ const putSale = async (saleId, body) => {
   };
 
 const deleteSale = async (saleId) => {
+  if (!saleId) return false;
+
   const query = 'DELETE FROM StoreManager.sales_products WHERE sale_id = ?';
   await connection.execute(query, [saleId]);
 };
